Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 1: Commit-Queue+1
> > 
> > (CLA confirmed.)
> > 
> > The level of arcania to get entropy on Windows is a little crazy :(
> > 
> > Obviously we don't have any builders nor consumers who exercise this because this is the first time that I'm aware that anyone has hit this. Therefore there's nothing to stop us regressing support for UWP in the future. Typically we are hesitant to support things that we can't support property but, I think this counts as "Windows", and we're supposed to work on Windows.
> > 
> > I'll leave this a little while because David or Ryan might have good ideas. Mostly I'm assuming from your email address that this is correct because I know nothing!
> 
> The context for this is https://crbug.com/boringssl/307. This CL is gated on UWP, so it won't directly run into Chrome's sandboxing needs, but I would like a clear answer as to what BCryptGenRandom with that flag actually does. If nothing else, that way we can document why we still have the RtlGenRandom codepath, rather than just sending everything through BCryptGenRandom.

I've got in touch with the owners of BCrypt.

Both RtlGenRandom and BCryptGenRandom end up using the same internal RNG infrastructure. 
RtlGenRandom is an older API. BCryptGenRandom is the public API for randomness on Windows. They suggested two options for replacing RtlGenRandom usage:

1. BCryptGenRandom( NULL, pbBuf, cbBuf, BCRYPT_USE_SYSTEM_PREFERRED_RNG )
    It is the code used by this CL. Which works on Win7 and later.
    This checks the system preferences, opens an RNG algorithm provider and provides the random bytes. The first call is slow, subsequent calls are much faster.

2. The slightly simpler method is to use
                BCryptGenRandom( BCRYPT_RNG_ALG_HANDLE, pbBuf, cbBuf, 0 )
The first parameter is a pseudo-handle (https://docs.microsoft.com/en-us/windows/win32/seccng/cng-algorithm-pseudo-handles) which behaves as an algorithm handle. 

The second solution is slightly faster, and conceptually simpler, but it is only available on Win10.

BCrypt* functions do add some overhead to handle the plug-in architecture, but once you’ve succeeded the first time the path from the call to the point you enter the actual RNG infrastructure code should be < 100 cycles.

Patch-set: 3