Update patch set 2
Patch Set 2:

Thanks for this. The BUILD file is generated so I'd need to apply this fix elsewhere but, first, can you confirm that this is a safe no-op for current Bazel versions? I.e. that can be deploy this immediately and not cause anyone to blow up?

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>