Keep the encryption state and encryption level in sync.
This is a little bit of cleanup along the way to fixing up the QUIC
callbacks. There is one TLS 1.3 state transition which doesn't carry an
encryption level: switching from 0-RTT keys back to unencrypted on an
HRR-based 0-RTT reject. The TCP code doesn't care about write_level and
the QUIC code is currently fine because we never "install" the 0-RTT
keys. However, the new encryption secrets API will hook in at install
time, so we should get this correct.

This also opens the door for DTLS 1.3, if we ever implement it, because
DTLS 1.3 will need to know which level it is to handle 0-RTT keys funny.
(Clients sending 0-RTT will briefly have handshake and 0-RTT write keys
active simultaneously.)

QUIC has the same property, but we can fudge it because only the caller
is aware of this.

Bug: 303
Change-Id: Ia76d787e1b96a058d9818948b6d9a051e8592207