Add a hook that allows X509_verify_cert() to always fail.
In order to flush out unwanted callers of X509_verify_cert(), this
commit adds a hook that allows the implementation to always return a
distinctive failure code, with access to the "real" implementation
controlled by a whitelist.

By default, the "real" implementation continues to be linked, and
there is no change in behavior.

Change-Id: I664cdcd6cabb0d2bbd9ab71cc1efd584fee04dea