Update patch set 4
Patch Set 4:

> I'm also seeing that BUILD.generated_tests.bzl is generated but no tests seem to exist.
> Any special reason for that?

It contains the source-sets that you would need in order to write test targets, but it's split out because, in our internal setup, the test configuration is pretty bespoke and not something suitable for the outside world. It would be quite possible to write the test targets in BUILD.toplevel, but nobody has ever requested it.

Patch-set: 4