Add SSL_SESSION_copy_without_early_data.
While we could store an extra boolean along with each session in the session
cache and then disable early data on a per-socket level, that causes
SSL_early_data_reason to report confusing values. Clearing it at the session
seems simpler. Since sessions are intended to be immutable, do this as a copy
operation.

Bug: chromium:1066623
Change-Id: I599b1559b696805e330ab5c2d61e4158440daef8