Update patch set 9
Patch Set 9:

I played around with this. This probably fixes Folly, but not for the reasons we thought. :-) Folly's problem is they do this unreasonable thing when they believe weak symbols aren't supported:
https://github.com/facebook/folly/blob/594f484b011d3f759a70b3173a4337690fae4651/folly/memory/detail/MallocImpl.cpp#L34-L46

If I stick void (*sdallocx)(void*, size_t, int) = NULL in another file, we still crash because a function pointer symbol with value NULL isn't the same as a weak function symbol that doesn't exist. So that means if we believe __attribute__(weak) works but Folly doesn't, we'll crash.

But we used to believe weak symbols were supported even on platforms where it wasn't. I believe now, as long as our weak symbol support check is a subset of Folly's, things are fine.

Still, what Folly is doing is really rude and, IMO, invalid. They should use inline namespaces so they aren't stomping on someone else's symbols in the fallback case.

Patch-set: 9