Clean up various EC inversion functions.
This fixes two issues. First, we have been lax about whether the low-level
inversion functions fail on zero input or output zero. Fix the documentation
and call the latter inv0 or inverse0 to match the terminology used in
draft-irtf-cfrg-hash-to-curve. (Although we may not actually need inv0 given
the optimization in D.2.)

This has no actual effect because the functions were only used in contexts
where the inputs were already guaranteed to be non-zero. Still, we should be
consistent here.

Second, ec_scalar_inv_montgomery and ec_scalar_inv_montgomery_vartime claim to
perform the same operation, but they do not. First, one computed inv0 and the
other computed inv (except only in some implementations, so fix it to be
consistent). Second, the former computes inverses in the Montgomery domain,
while the latter converts to the Montgomery domain and then inverts. Rename it
to ec_scalar_to_montgomery_inv_vartime, which is... questionably understandable
but at least looks different.

Change-Id: I9b4829ce5013bdb9528078a093f41b1b158df265