Optimizing ECDSA P-384 signing
This commit optimizes ECDSA P-384 signing. The specific implementation improves the performance on X86-64 platforms with assembly enabled. It uses some of the
methods that are explained in
    S. Gueron, V. Krasnov, "Fast Prime Field Elliptic Curve Cryptography with 256 Bit Primes", Journal of Cryptographic Engineering 5:141-151 (2015)
and are used in the P-256 code.

The code speeds up ECDSA P-384 signing by a factor of x3.5 (= 2335/666 operations per second) when tested on a Dell XPS 13 7390 2-in-1 laptop. This platform is equipped with the 10th generation Intel(c) Core(TM) processor (microarchitecture codename "Ice Lake"[ICL]). The benchmark compiled the code using gcc-9 on Ubuntu 18.0.4 LTS, and the performance was measured with Intel(c) Turbo Boost Technology disabled.

This commit is due to Nir Drucker (ndrucker@amazon.com) and Shay Gueron (gueron@amazon.com) from the AWS Crpyographic Algorithm Team.

Details: the code
- Modifies the make_p256-x86_64-table.go script to generate a table for the P-384 curve (p384-x86_64-table.h) too.
- Merges common utility functions for p-256 and p-384 in p256_p384_common-x86_64.h.
- New mul_base function for P-384 is defined in p384.c

For simplicity, this commit avoids any assembly code usage. Further speed up could be obtained by optimizing lower level assembly functions, but we chose to skip this option (here) for simplicity.

The code runs in constant-time under the assumption that the following functions also run in constant-time: "ec_GFp_mont_add" and "ec_felem_neg".

Change-Id: I76d3991a39e0368b53d0c7c5ac39c81acb9d9d02