Update patch set 4
Patch Set 4:

> Err, to the TRUST_TOKEN_ISSUER_new and CLIENT version mentioning that batchsize should be 16-bit?

Ah yeah, we should probably do that. Alternatively, would it make sense to simply clamp them to 0xffff? I don't entirely understand how the batchsize parameter works. Is it a parameter to the system, such that, if we see a big batchsize it is an config error somewhere, or is it just some common place to ensure you don't ask for too many tokens, in which case we can always freely issue fewer tokens than requested.

(E.g. if the client requests more than max_batchsize tokens, we seem to just ignore that, rather than treat it as an error.)

Patch-set: 4
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>