Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> > Probably make most of the API use key_id instead of public_metadata, except for the SRR.
> 
> Not sure I follow. I thought, from the perspective of the public functions, at least, the abstraction was public_metadata, and key_id was an internal thing. Maybe I'm misunderstanding?
Should've been on the pmbtokens file, though it might just be fixing that one line. For the sake of the PBMTokens API, everything is key_id (it doesn't know about the concept of public_metadata). The only public_metadata is the issue and SRR calls.

Patch-set: 4
Reviewer: Gerrit User 5955 <5955@66b711a4-aa9b-3dac-9062-916540830212>