Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (1 comment)
> > 
> > > Probably make most of the API use key_id instead of public_metadata, except for the SRR.
> > 
> > Not sure I follow. I thought, from the perspective of the public functions, at least, the abstraction was public_metadata, and key_id was an internal thing. Maybe I'm misunderstanding?
> Should've been on the pmbtokens file, though it might just be fixing that one line. For the sake of the PBMTokens API, everything is key_id (it doesn't know about the concept of public_metadata). The only public_metadata is the issue and SRR calls.

Ah. I think that's true now? I tried to take the key_id out of pmbtokens.c entirely, but I didn't quite succeed because the key_id is in the TRUST_TOKEN object. Then I had second thoughts as to whether that was actually correct. Then I waffled for a bit and decided to upload the darn thing. :-) (Thoughts?)

Patch-set: 4