Still query getauxval if reading /proc/cpuinfo fails.
If BoringSSL is used in a sandbox without /proc/cpuinfo, we will
silently act as if the CPU is missing capabilities, even though
getauxval may be available. We use /proc/cpuinfo to work around a
missing AT_HWCAP2 and ignore a particular broken CPU.

The latter is now vanishingly rare (even missing AT_HWCAP2 has largely
dropped off), so instead proceed with getauxval. This makes the /proc
paths largely optional.

Change-Id: Ib198c4f78ccdae874d55669b6a7508dfbeac0f44