Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > The design of "require all Transport Parameters to byte-for-byte match" is broken. One of the Transport Parameters is the stateless reset token; this is required to be different per connection, so necessarily the byte-for-byte match will fail.
> 
> Hrmf. I guess we should have a callback of some sort then? Or would it be easier for BoringSSL to know about QUIC transport parameters internally? (Guessing not since you will probably be defining them all over the place.) Or maybe you have the caller extract a distinct "context" from the transport params? Not sure which would be more fuss.
> 
> Would we want a similar picture for HTTP/3?

I don't think BoringSSL should get in the business of parsing or otherwise handling transport parameters that involves anything more than treating it as a bag of bytes. I think the simplest fix is to ignore the transport params and use only the quic early data context. I would accompany such a change with requiring SSL_set_quic_early_data_context to be called (with a non-empty value) to allow early data and a documentation change that both transport params and application context need to be covered in the data provided there.

Patch-set: 6