Update patch set 6
Patch Set 6:

> Mostly out of curiosity, if the stateless reject token can freely change from 0-RTT to 1-RTT, how does that work? Is stateless reject just not applicable yet until the ServerHello so we effectively don't remember it?

The short answer is section 7.4.1 of quic-transport ("Values of Transport Parameters for 0-RTT") states "A client MUST NOT use remembered values for the following parameters: ack_delay_exponent, max_ack_delay, initial_source_connection_id, original_destination_connection_id, preferred_address, retry_source_connection_id, and stateless_reset_token. The client MUST use the server's new values in the handshake instead, and absent new values from the server, the default value."

The long answer: The stateless reset token is a mechanism for the server to close a connection when it has lost all other connection state. (If a server->client packet matches this opaque blob, the client uses that as the signal that the server just killed the connection.) To prevent injection attacks, the stateless reset token must be dependent on the Connection ID of the connection it is used on. Since it's dependent on the CID, there's no way for the stateless reset token from the previous connection to apply on the new connection. Also, since the purpose of the stateless reset token is to let the server tear down a connection even though it's missing connection state, such a facility isn't needed if the server hasn't sent anything on the connection yet.

Patch-set: 6