Update patch set 1
Patch Set 1:

We generally only cross-compile to Android for AArch64, for which the BUILDING.md instructions seem to work. Did you hit issues with that, or are you trying to cross-compile to a non-Android Linux target? If the latter could you expand the commit message to explain the impact of this?

Patch-set: 1
CC: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>