Update patch set 1
Patch Set 1: Code-Review+2

Will leave for a little while to see if David has thoughts, but seems good.

R.e. SVE2:

I'm not sure what you're aiming at but, if it's a future huge change that includes lots of SVE2 assembly, here's some things to keep in mind:

We're majority focused on Google's needs. I assume that SVE2 will come to phones and that would take care of the motivation but, if it's server-chip only or such, then the motivation may be lacking.

We need to be able to test code. We don't want a situation where there's a lot of dark code that our bots can't test, that we don't have the hardware for, and that blows up for some Android engineer bringing up a new device. Intel have SDE, a valgrind-like simulator, that we can use for this, but we don't have anything like that for ARM.

Intrinsics have a lot of benefits. If not applicable here then assembly code needs to unwind correctly (generally using CFI annotations).

Code size is important on ARM.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+2