[perl][armv8] Support branch-protection in scripts
This change adds optional support for
 - Armv8.3-A Pointer Authentication and
 - Armv8.5-A Branch Target Identification (BTI)
features to the perl scripts.

Both of these features can be enabled with additional compiler flags
and are off by default.

This change adds BTI landing pads even to assembly functions which are
likely to be directly called only. In these cases landing pads might be
superfluous.

Change-Id: I4335f92e2ccc8e209c7d68a0a79f1acdf3aeb791