Use |crypto_word_t| and |size_t| more consistently in ECC scalar recoding.
Use |crypto_word_t| as the type for secret values in scalar recoding. Use
|size_t| as the type of array indexes in scalar recoding. Use explicit
casts where a larger type is (losslessly) truncated to a smaller type. With
this change, |uint64_t| is no longer used in the p256.c when building in
32-bit mode, |unsigned| is not used in any of the affected modules, and
|uint8_t| and |char| are no longer used for secret values in the ECC
recoding.

This is part of a bigger project to minimize the use of types other than
|crypto_word_t| for secret values. This is also part of a larger project
make the ECC code more consistent.

Avoid changing the loop indexing in the P-256 scalar multiplication from |int|
to |size_t|. The P-224 code does use |size_t| but it is less clear than the
P-256 code where |i - 1| results in a negative/underflowed value when |i| is
zero.

Remove unneeded pointer casts during point selection.

Change-Id: Icec7b907d004d6daca0e78d2f74438995f1ff131