Update patch set 4
Patch Set 4:

I restored the pointer casts that an earlier version of the CL had removed. It seems like clang and MSVC don't require them but some versions of GCC do. Maybe that's still a good change to make, but it's unrelated to the changes here and so they should be in a separate CL.

Patch-set: 4